Implement AbstractElementManipulator of GnStringExprImpl to not crash on rename
Currently, when a file that is referenced in GN is renamed, the IDE
crashes. This is because AbstractElementManipulator is not implemented
for GnStringExprImpl, which is the PSI element that represents a string
literal in GN.
This commit implements the method, which allows IntelliJ to rename
strings in GN files. However, this is a workaround, as the string
literal is not actually renamed in the GN file. Future work will be
needed to implement this functionality.